Windows build warning and Makefile update
.\hack.c(2657): warning C4389: '!=': signed/unsigned mismatch
Visual Studio autodetection update for Makefile.msc

(cherry picked from commit 254c6c55f5b163b343d08bd915334e8f1f3b29a7)

+++ src/hack.c
@@ -2815,7 +2815,7 @@ check_special_room(boolean newlev)
                     if (DEADMONSTER(mtmp))
                         continue;
                     if (!isok(mtmp->mx,mtmp->my)
-                        || roomno != levl[mtmp->mx][mtmp->my].roomno)
+                        || roomno != (int) levl[mtmp->mx][mtmp->my].roomno)
                         continue;
                     if (!Stealth && !rn2(3)) {
                         if (mtmp->msleeping && canseemon(mtmp))
